# proxy/docs/swagger.yaml
definitions:
  main.Address:
    properties:
      city:
        type: string
      geo_lat:
        type: string
      geo_lon:
        type: string
      house:
        type: string
      street:
        type: string
    type: object
  main.ErrorResponse:
    description: Ошибка, возникающая при обработке запроса
    properties:
      "200":
        type: string
      "400":
        type: string
      "500":
        type: string
    type: object
  main.GeocodeRequest:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  main.RequestAddressSearch:
    properties:
      query:
        type: string
    type: object
  main.ResponseAddress:
    properties:
      suggestions:
        items:
          properties:
            data:
              $ref: '#/definitions/main.Address'
          type: object
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Этот эндпоинт позволяет получить адрес по геоданным
  title: Address API
  version: "1.0"
paths:
  /api/address/geocode:
    post:
      description: This endpoint allows you to get geo coordinates by address
      parameters:
      - description: Address search lat lng
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/main.GeocodeRequest'
      responses:
        "200":
          description: Успешное выполнение
          schema:
            $ref: '#/definitions/main.ResponseAddress'
        "400":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Ошибка подключения к серверу
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get Geo Coordinates
  /api/address/search:
    post:
      description: This endpoint allows you to get geo coordinates by address
      parameters:
      - description: Address search query
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/main.RequestAddressSearch'
      responses:
        "200":
          description: Успешное выполнение
          schema:
            $ref: '#/definitions/main.ResponseAddress'
        "400":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Ошибка подключения к серверу
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get Geo Coordinates
swagger: "2.0"